runs intake while shooting
the intake runs with the shooter now in case any power cells get stuck
not all the way fed in. also changed the distance for the movement at
the start and gets the start angle every time the command runs instead
of just instantiated so it can be run multiple times.