Finished autonomous on A robot
Moved the code in RotateToAngle.java that determines which direction to
move in. Since it ran every execute loop, it would change direction
every loop and cause the robot to oscillate at its target.

Also changed the equation in the outtake subsystem to tune the shooter
values to be more accurate from more distances.